+++ packages/xchain-util/src/chain.ts
@@ -7,6 +7,7 @@ export enum Chain {
   Polkadot = 'POLKA',
   BitcoinCash = 'BCH',
   Litecoin = 'LTC',
+  Terra = 'LUNA',
 }
 
 export const BNBChain = Chain.Binance
@@ -17,6 +18,7 @@ export const CosmosChain = Chain.Cosmos
 export const PolkadotChain = Chain.Polkadot
 export const BCHChain = Chain.BitcoinCash
 export const LTCChain = Chain.Litecoin
+export const LUNAChain = Chain.Terra
 
 /**
  * Type guard to check whether string  is based on type `Chain`
@@ -46,5 +48,6 @@ export const chainToString: ((chainId: Chain) => string) & Record<Chain, string>
     [Chain.Binance]: 'Binance Chain',
     [Chain.Cosmos]: 'Cosmos',
     [Chain.Polkadot]: 'Polkadot',
+    [Chain.Terra]: 'Terra',
   },
 )